refactor(html5): extract shared getSDK util + harden Panopto URL validation
