+++ bigbluebutton-html5/imports/ui/components/external-video-player/custom-players/arc-player.jsx
@@ -1,49 +1,11 @@
-import loadScript from 'load-script';
 import React, { Component } from 'react'
+import getSDK from './get-sdk';
 
 const MATCH_URL = new RegExp("https?:\/\/(.*)(instructuremedia.com)(\/embed)?\/([-abcdef0-9]+)");
 
 const SDK_URL = 'https://files.instructuremedia.com/instructure-media-script/instructure-media-1.1.0.js';
 
 const EMBED_PATH = "/embed/";
-
-// Util function to load an external SDK or return the SDK if it is already loaded
-// From https://github.com/CookPete/react-player/blob/master/src/utils.js
-const resolves = {};
-export function getSDK (url, sdkGlobal, sdkReady = null, isLoaded = () => true, fetchScript = loadScript) {
-  if (window[sdkGlobal] && isLoaded(window[sdkGlobal])) {
-    return Promise.resolve(window[sdkGlobal])
-  }
-  return new Promise((resolve, reject) => {
-    // If we are already loading the SDK, add the resolve
-    // function to the existing array of resolve functions
-    if (resolves[url]) {
-      resolves[url].push(resolve);
-      return
-    }
-    resolves[url] = [resolve];
-    const onLoaded = sdk => {
-      // When loaded, resolve all pending promises
-      resolves[url].forEach(resolve => resolve(sdk))
-    };
-    if (sdkReady) {
-      const previousOnReady = window[sdkReady];
-      window[sdkReady] = function () {
-        if (previousOnReady) previousOnReady();
-        onLoaded(window[sdkGlobal])
-      }
-    }
-    fetchScript(url, err => {
-      if (err) {
-        reject(err);
-      }
-      window[sdkGlobal] = url;
-      if (!sdkReady) {
-        onLoaded(window[sdkGlobal])
-      }
-    })
-  })
-}
 
 export class ArcPlayer extends Component {
   static displayName = 'ArcPlayer';
