+++ bigbluebutton-html5/imports/ui/components/external-video-player/custom-players/peertube.jsx
@@ -1,48 +1,10 @@
-import loadScript from 'load-script';
 import React, { Component } from 'react'
+import getSDK from './get-sdk';
 
 //To work with PeerTube >=v3.3 URL patterns
 const MATCH_URL = new RegExp("(https?)://(.*)(/videos/watch/|/w/)(.*)");
 
 const SDK_URL = 'https://unpkg.com/@peertube/embed-api@0.0.4/build/player.min.js';
-
-// Util function to load an external SDK or return the SDK if it is already loaded
-// From https://github.com/CookPete/react-player/blob/master/src/utils.js
-const resolves = {};
-export function getSDK (url, sdkGlobal, sdkReady = null, isLoaded = () => true, fetchScript = loadScript) {
-  if (window[sdkGlobal] && isLoaded(window[sdkGlobal])) {
-    return Promise.resolve(window[sdkGlobal])
-  }
-  return new Promise((resolve, reject) => {
-    // If we are already loading the SDK, add the resolve
-    // function to the existing array of resolve functions
-    if (resolves[url]) {
-      resolves[url].push(resolve);
-      return
-    }
-    resolves[url] = [resolve];
-    const onLoaded = sdk => {
-      // When loaded, resolve all pending promises
-      resolves[url].forEach(resolve => resolve(sdk))
-    };
-    if (sdkReady) {
-      const previousOnReady = window[sdkReady];
-      window[sdkReady] = function () {
-        if (previousOnReady) previousOnReady();
-        onLoaded(window[sdkGlobal])
-      }
-    }
-    fetchScript(url, err => {
-      if (err) {
-        reject(err);
-      }
-      window[sdkGlobal] = url;
-      if (!sdkReady) {
-        onLoaded(window[sdkGlobal])
-      }
-    })
-  })
-}
 
 export class PeerTubePlayer extends Component {
   static displayName = 'PeerTubePlayer';
